refactor(renderer): split the preset views out of camera-framing.ts
camera-framing.ts landed at 415 lines. The repo's rule says ~400 non-generated
lines, and its only enforced implementation of that same sentence — Rust's
module_size_ratchet — counts raw lines (`s.lines().count()`) against a hard
400, so 415 is over on the measure the repo actually uses, comments included.

The seam is by subject, not by line count: camera-framing.ts keeps the FREE
pickers, which inherit the current view direction and answer 'put the camera on
this thing' (point / bounds / extents). camera-preset-view.ts takes the
ViewCube's NAMED directions, where the direction comes from the face plus the
caller's rotation cycle and the box has to clear its own admission check first.
They share only box arithmetic, so centerOf/maxExtentOf are exported rather than
copied.

Pure move: no test imports either module directly (both are reached through
CameraAnimator), no guard changes hands, and every leaf count is unchanged.
